Lisätty kalorilaskuri ja sen sisälogiikka.

+++ Calo4U_Sisa/Kayttaja.cs
@@ -9,8 +9,8 @@ namespace Calo4U_Sisa
     internal class Kayttaja
     {
         public string Nimi {  get; set; }
-        public  int PvaKalorit {  get; set; }
-        public int HaeKayttajanKalorit()
+        public  double PvaKalorit {  get; set; }
+        public double HaeKayttajanKalorit()
         {
             return PvaKalorit;
         }
